feat：报销单 空指针问题

+++ src/main/java/com/jtyjy/finance/manager/service/impl/BudgetExtractCommissionApplicationLogServiceImpl.java
@@ -153,7 +153,9 @@ public class BudgetExtractCommissionApplicationLogServiceImpl extends ServiceImp
         if(Objects.equals(logStatus, LogStatusEnum.REJECT.getCode())) {
             if (application.getReimbursementId() != null) {
                 BudgetReimbursementorder reimbursementOrder = reimburseService.getById(application.getReimbursementId());
-                reimburseService.removeById(reimbursementOrder.getId());
+                if (reimbursementOrder!=null) {
+                    reimburseService.removeById(reimbursementOrder.getId());
+                }
             }
             BudgetExtractsum budgetExtractsum = extractSumMapper.selectById(sumId);
             budgetExtractsum.setStatus(ExtractStatusEnum.RETURN.getType());
